added ipv6 support & config file

+++ ip2host.php
@@ -1,22 +1,31 @@
 <?php
+require("config.php");
 require("namecom.php");
 $api = new NameComApi();
-$api->login('NAME_COM_API_USERNAME', 'NAME_COM_API_KEY');
+$api->login($api_username, $api_key);
 
 $sub_domain = $_GET['sub_domain'];
 $super_domain = $_GET['super_domain'];
+
+if (!isset($super_domain)) {
+	die(json_encode(array("response"=>"You must specify a super domain.")));
+}
+
 $ip = ((@$_GET['ip']) ? $_GET['ip'] : $_SERVER['REMOTE_ADDR']);
 $ttl = ((@$_GET['ttl']) ? $_GET['ttl'] : 300); // default 5 minute time to live
 
 $response = $api->get_domain($super_domain);
+
 if ($response->result->code != 100) {
         $a = array("response"=>"You probably don't own the domain.","code"=>$response->result->code);
 }
 
 $response = $api->list_dns_records($super_domain);
+
 if ($response->result->code != 100) {
         $a = array("response"=>"An unknown error has occured when querying the superdomain.","code"=>$response->result->code);
 }
+
 foreach ($response->records as $key => $val) {
         if ($val->name == $sub_domain.".".$super_domain) { // delete entry
                 $_response = $api->delete_dns_record($super_domain, $val->record_id);
@@ -25,11 +34,23 @@ foreach ($response->records as $key => $val) {
                 }
         }
 }
-$response = $api->create_dns_record($super_domain, $sub_domain, 'A', $ip, $ttl);
+
+if (!filter_var($ip, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6) === false) {
+        // $ip is valid IPv6
+        $record_type = "AAAA";
+} else {
+        $record_type = "A";
+}
+
+$response = $api->create_dns_record($super_domain, $sub_domain, $record_type, $ip, $ttl);
+
 if ($response->result->code != 100) {
         $a = array("response"=>"An unknown error has occured when updating the record.","code"=>$response->result->code);
 } else {
         $a = array("response"=>"IP updated successfully.","code"=>$response->result->code);
 }
+
+echo json_encode($a);
+
 $api->logout();
 ?>
